AshPyroComponents.Components.Form (AshPyroComponents v0.1.0)

A component that auto-renders forms for Ash resources.

Overridable Component Attributes

You can customize the components in this module by configuring overrides.

The components in this module support the following overridable attributes:

ash_form/1

:autocomplete :string (required)

:class :css_classes

:actions_class :css_classes

render_field/1

:field_group_class :css_classes

:field_group_label_class :css_classes

Components

ash_form(assigns)

Renders an Ash form.

Attributes

overrides (:list) - Manually set the overrides for this component (instead of config/default). Defaults to nil.

action_info (:any) - Defaults to :unassigned.

pyro_form (:any) - Defaults to :unassigned.

as (:any) - the server side parameter to collect all input under. Defaults to nil.

for (:map) (required) - the datastructure for the form.

resource (:atom) (required) - the resource of the form.

actor (:map) - the actor to be passed to actions. Defaults to nil.

tz (:string) - timezone. Defaults to "Etc/UTC".

autocomplete (:string) - (overridable, required).

class (:any) - (overridable, :css_classes).

actions_class (:any) - (overridable, :css_classes).

Global attributes are accepted. the arbitrary HTML attributes to apply to the form tag. Supports all globals plus: ["name", "rel", "action", "enctype", "method", "novalidate", "target"].

Slots

actions - extra form actions.